Update Pacman Game using C
Pacman is the most famous 2D game played. Pacman is a single-player game. The rules of the game are mentioned below:

This a single-player game
need to collect dots to complete the level
If all dots are collected level is completed


The Functionality of the game is :

Play the game
Calculate the score
Maintain the top score